feat: add configurable path exclusion filtering to ado and github ingestors
