+++ plugins/ai-assistant-common/src/utils/path-filter.ts
@@ -1,43 +1,11 @@
 import { minimatch } from 'minimatch';
 
 /**
- * Default path exclusion patterns for common build artifacts and dependencies
- */
-export const DEFAULT_PATH_EXCLUSIONS = [
-  'node_modules/**',
-  '.git/**',
-  '.github/workflows/**',
-  'dist/**',
-  'build/**',
-  'target/**',
-  'vendor/**',
-  '.next/**',
-  '.nuxt/**',
-  'coverage/**',
-  '.nyc_output/**',
-  'out/**',
-  'bin/**',
-  'obj/**',
-  '.vscode/**',
-  '.idea/**',
-  '*.min.js',
-  '*.min.css',
-  '.env',
-  '.env.*',
-  '*.log',
-  'tmp/**',
-  'temp/**',
-  '.cache/**',
-  'bower_components/**',
-  'jspm_packages/**',
-];
-
-/**
  * Options for path filtering
  */
 export interface PathFilterOptions {
   /** Array of glob patterns to exclude from processing */
-  exclusionPatterns?: string[];
+  exclusionPatterns: string[];
   /** Whether to use case-insensitive matching (default: true) */
   caseInsensitive?: boolean;
 }
@@ -46,11 +14,8 @@ export interface PathFilterOptions {
  * Creates a path filter function that can determine if a file path should be excluded
  * based on configured glob patterns
  */
-export function createPathFilter(options: PathFilterOptions = {}) {
-  const {
-    exclusionPatterns = DEFAULT_PATH_EXCLUSIONS,
-    caseInsensitive = true,
-  } = options;
+export function createPathFilter(options: PathFilterOptions) {
+  const { exclusionPatterns, caseInsensitive = true } = options;
 
   /**
    * Tests if a file path should be excluded based on configured patterns
